Turned out trickier than expected. Had to initialize a client container containing the splash and the portfolio slot in order to enable smooth-scrolling behavior.
