+++ src/app/@portfolio/(.)engineering/page.tsx
@@ -1,0 +1,10 @@
+import EngineeringPage from "@/app/engineering/page";
+
+export default function EngineeringPortfolio() {
+
+  return (
+    <>
+      <EngineeringPage />
+    </>
+  )
+}
