+++ src/app/@portfolio/(.)music/page.tsx
@@ -1,0 +1,10 @@
+import MusicPage from "@/app/music/page"
+
+export default function MusicPortfolio() {
+
+  return (
+    <>
+      <MusicPage />
+    </>
+  )
+}
