+++ src/app/@portfolio/_layout.tsx
@@ -1,0 +1,8 @@
+export default function PortfolioLayout({ children }: { children: React.ReactNode }) {
+
+  return (
+    <>
+      {children}
+    </>
+  )
+}
