--- src/app/@portfolio/page.tsx
@@ -1,8 +1,0 @@
-export default function PortfolioSlot() {
-
-  return (
-    <>
-      Portfolio Slot
-    </>
-  )
-}
