+++ src/app/_components/Navbar.tsx
@@ -14,8 +14,6 @@ export default function Navbar() {
   const darkModeButton = useRef<HTMLDivElement>(null);
   const { darkMode, toggleDarkMode } = useContext(DarkModeContext)
 
-  const [showHome, setShowHome] = useState<boolean>(false);
-
   const toggleModeButton = () => {
     if (!darkMode) {
       darkModeButton.current?.classList.remove("bg-yellow-500", "-translate-x-2");
@@ -37,16 +35,11 @@ export default function Navbar() {
     <div id="navbar"
       className="absolute w-full">
       <div className="relative size-10 md:size-12 left-5 top-5 dark:text-stone-100">
-        <button className="w-full" onClick={() => setShowHome(!showHome)}>
-          {homeIcon}
-        </button>
-        {showHome && <>
-          <div className="absolute w-48 left-11 top-0 bg-slate-500/50 flex justify-center">
-            <p className="text-sm">
-              Home is where my dog is.
-            </p>
+        <Link href="/">
+          <div className="w-full">
+            {homeIcon}
           </div>
-        </>}
+        </Link>
       </div>
       <div id="dark-mode-toggle-box"
         className="absolute right-8 md:right-10 lg:right-14 top-6 w-8 h-4">
