+++ src/app/engineering/_components/ProjectMediaDisplay.tsx
