+++ src/app/engineering/_components/TechStackDisplay.tsx
@@ -1,10 +1,8 @@
-
 interface TechStackDisplayProps {
   className: string,
   projectName: string,
   techStack: string[],
 }
-
 
 export default function TechStackDisplay(props: TechStackDisplayProps) {
