+++ src/app/engineering/page.tsx
@@ -1,0 +1,13 @@
+import PortfolioHeader from "../_components/PortfolioHeader";
+import EngineeringContainer from "./_containers/EngineeringContainer";
+
+export default function EngineeringPage() {
+
+  return (
+    <>
+      <EngineeringContainer>
+        <PortfolioHeader category="engineering" />
+      </EngineeringContainer>
+    </>
+  )
+}
